update: add `nvim-tundra` colorscheme and disable `catppuccin`
This is my personally-created colour scheme after the catppuccin scheme
changed a lot of its colours and themes (in my opinion) for the worst.

You can find my colorscheme
[here](https://github.com/sam4llis/nvim-tundra).
It aims to replicate a lot of the niceties of the original catppuccin
theme, though I have also changed colours and highlights for certain
elements.